Remove Telegram agent bc mimemagic/httmultiparty
The Problem:
============
The Telegram agent requires the unmaintained httmultiparty, which in
turn depends on mimemagic, which [has system dependencies](https://www.theregister.com/2021/03/29/rails_mime_fix/).

The Solution:
=============
I don't use Telegram, so remove the agent and its dependency. Note that
I've not checked the Gemfile for other things that could be removed.